requirements.txt
ascii_magic
pillow
requests
geocoder
python-dotenv
